# Security: Battle-tested CodeQL config + ReDoS fix + array memoization #777