… build/hooks/attribution-handler.js